Update ArcGIS_Pro_Notebook_Map_Automation.py
Updated to the 100% AUTOMATED version of this script. Analyst no longer has to choose one of the buffered or non-buffered output maps, and the script runs twice as fast because it doesn't have to output two of everything. There are also now no edge cases in which a map must be made manually.